move calculator under metrics folder
