+++ src/main/java/org/opensearch/searchrelevance/metrics/EvaluationMetrics.java
@@ -7,13 +7,13 @@
  */
 package org.opensearch.searchrelevance.metrics;
 
-import static org.opensearch.searchrelevance.calculator.Evaluation.METRICS_MEAN_AVERAGE_PRECISION;
-import static org.opensearch.searchrelevance.calculator.Evaluation.METRICS_NORMALIZED_DISCOUNTED_CUMULATIVE_GAIN;
-import static org.opensearch.searchrelevance.calculator.Evaluation.METRICS_PRECISION_AT_10;
-import static org.opensearch.searchrelevance.calculator.Evaluation.METRICS_PRECISION_AT_5;
-import static org.opensearch.searchrelevance.calculator.Evaluation.calculateMAP;
-import static org.opensearch.searchrelevance.calculator.Evaluation.calculateNDCG;
-import static org.opensearch.searchrelevance.calculator.Evaluation.calculatePrecisionAtK;
+import static org.opensearch.searchrelevance.metrics.calculator.Evaluation.METRICS_MEAN_AVERAGE_PRECISION;
+import static org.opensearch.searchrelevance.metrics.calculator.Evaluation.METRICS_NORMALIZED_DISCOUNTED_CUMULATIVE_GAIN;
+import static org.opensearch.searchrelevance.metrics.calculator.Evaluation.METRICS_PRECISION_AT_10;
+import static org.opensearch.searchrelevance.metrics.calculator.Evaluation.METRICS_PRECISION_AT_5;
+import static org.opensearch.searchrelevance.metrics.calculator.Evaluation.calculateMAP;
+import static org.opensearch.searchrelevance.metrics.calculator.Evaluation.calculateNDCG;
+import static org.opensearch.searchrelevance.metrics.calculator.Evaluation.calculatePrecisionAtK;
 
 import java.util.HashMap;
 import java.util.List;
@@ -26,40 +26,13 @@ public class EvaluationMetrics {
 
     /**
      * calculate evaluation metrics with evaluation calculators.
-     * @param indexToDocIdMap - indexToDocIdMap that has search configuration index to doc id list
-     * e.g:
-     * // indexToDocIdMap
-     * {
-     *     "0": ["docId2"],
-     *     "1": ["docId2"]
-     *     "2": ["docId1","docId2","docId3"]
-     * }
-     * // judgments
-     * {
-     *     "docId1": 0.9,
-     *     "docId2": 0.85,
-     *     "docId3": 0.8
-     * }
-     * // evaluationMetrics
-     * {
-     *     "precision": 0.85   // please add more metrics here
-     * }
      */
-    public static Map<String, Map<String, Double>> calculateEvaluationMetrics(
-        Map<String, Object> indexToDocIdMap,
-        Map<String, Double> judgments
-    ) {
-        Map<String, Map<String, Double>> evaluationMetrics = new HashMap<>();
-        for (int i = 0; i < indexToDocIdMap.size(); i++) {
-            String searchConfigIndex = String.valueOf(i);
-            List<String> docIds = (List<String>) indexToDocIdMap.get(searchConfigIndex);
-            Map<String, Double> currSearchConfigMetrics = new HashMap<>();
-            currSearchConfigMetrics.put(METRICS_PRECISION_AT_5, calculatePrecisionAtK(docIds, judgments, 5));
-            currSearchConfigMetrics.put(METRICS_PRECISION_AT_10, calculatePrecisionAtK(docIds, judgments, 10));
-            currSearchConfigMetrics.put(METRICS_MEAN_AVERAGE_PRECISION, calculateMAP(docIds, judgments));
-            currSearchConfigMetrics.put(METRICS_NORMALIZED_DISCOUNTED_CUMULATIVE_GAIN, calculateNDCG(docIds, judgments));
-            evaluationMetrics.put(searchConfigIndex, currSearchConfigMetrics);
-        }
-        return evaluationMetrics;
+    public static Map<String, String> calculateEvaluationMetrics(List<String> docIds, Map<String, String> judgments) {
+        Map<String, String> currSearchConfigMetrics = new HashMap<>();
+        currSearchConfigMetrics.put(METRICS_PRECISION_AT_5, String.valueOf(calculatePrecisionAtK(docIds, judgments, 5)));
+        currSearchConfigMetrics.put(METRICS_PRECISION_AT_10, String.valueOf(calculatePrecisionAtK(docIds, judgments, 10)));
+        currSearchConfigMetrics.put(METRICS_MEAN_AVERAGE_PRECISION, String.valueOf(calculateMAP(docIds, judgments)));
+        currSearchConfigMetrics.put(METRICS_NORMALIZED_DISCOUNTED_CUMULATIVE_GAIN, String.valueOf(calculateNDCG(docIds, judgments)));
+        return currSearchConfigMetrics;
     }
 }
