+++ src/main/java/org/opensearch/searchrelevance/metrics/calculator/Evaluation.java
@@ -5,7 +5,7 @@
  * this file be licensed under the Apache-2.0 license or a
  * compatible open source license.
  */
-package org.opensearch.searchrelevance.calculator;
+package org.opensearch.searchrelevance.metrics.calculator;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -21,13 +21,13 @@ public class Evaluation {
     /**
      * Precision@K - measures precision at a specific rank k
      */
-    public static double calculatePrecisionAtK(List<String> docIds, Map<String, Double> judgmentScores, int k) {
+    public static double calculatePrecisionAtK(List<String> docIds, Map<String, String> judgmentScores, int k) {
         int relevantCount = 0;
         int count = 0;
 
         for (String docId : docIds) {
             if (count >= k) break;
-            if (judgmentScores.containsKey(docId) && judgmentScores.get(docId) > 0) {
+            if (judgmentScores.containsKey(docId) && Double.valueOf(judgmentScores.get(docId)) > 0) {
                 relevantCount++;
             }
             count++;
@@ -40,13 +40,13 @@ public class Evaluation {
     /**
      * Mean Average Precision (MAP)
      */
-    public static double calculateMAP(List<String> docIds, Map<String, Double> judgmentScores) {
+    public static double calculateMAP(List<String> docIds, Map<String, String> judgmentScores) {
         double sum = 0.0;
         int relevantCount = 0;
 
         for (int i = 0; i < docIds.size(); i++) {
             String docId = docIds.get(i);
-            if (judgmentScores.containsKey(docId) && judgmentScores.get(docId) > 0) {
+            if (judgmentScores.containsKey(docId) && Double.valueOf(judgmentScores.get(docId)) > 0) {
                 relevantCount++;
                 sum += (double) relevantCount / (i + 1);
             }
@@ -59,14 +59,14 @@ public class Evaluation {
     /**
      * Normalized Discounted Cumulative Gain (NDCG)
      */
-    public static double calculateNDCG(List<String> docIds, Map<String, Double> judgmentScores) {
+    public static double calculateNDCG(List<String> docIds, Map<String, String> judgmentScores) {
         double dcg = 0.0;
         double idcg = calculateIDCG(docIds, judgmentScores);
 
         for (int i = 0; i < docIds.size(); i++) {
             String docId = docIds.get(i);
             if (judgmentScores.containsKey(docId)) {
-                double relevance = judgmentScores.get(docId);
+                double relevance = Double.valueOf(judgmentScores.get(docId));
                 dcg += (Math.pow(2, relevance) - 1) / (Math.log(i + 2) / Math.log(2));
             }
         }
@@ -75,11 +75,11 @@ public class Evaluation {
         return Math.round(ndcg * 100.0) / 100.0;
     }
 
-    private static double calculateIDCG(List<String> docIds, Map<String, Double> judgmentScores) {
+    private static double calculateIDCG(List<String> docIds, Map<String, String> judgmentScores) {
         List<Double> relevanceScores = new ArrayList<>();
         for (String docId : docIds) {
             if (judgmentScores.containsKey(docId)) {
-                relevanceScores.add(judgmentScores.get(docId));
+                relevanceScores.add(Double.valueOf(judgmentScores.get(docId)));
             }
         }
