+++ src/main/java/org/opensearch/searchrelevance/metrics/calculator/PairComparison.java
@@ -5,7 +5,7 @@
  * this file be licensed under the Apache-2.0 license or a
  * compatible open source license.
  */
-package org.opensearch.searchrelevance.calculator;
+package org.opensearch.searchrelevance.metrics.calculator;
 
 import java.util.HashMap;
 import java.util.HashSet;
